backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.18
openai==1.3.7
transformers==4.35.2
torch==2.1.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
asyncpg==0.29.0
